Send speech input from options tab to other scripts
This is a trusted source, meaning that once the user gives permission
once, it is always given, so the options tab can listen continuously to
user input. This is necessary as websites that are not trusted (ie http
as opposed to https) have to ask every time microphone access is
requested and speech recognition times out after ~10 seconds of no
speech.