Make focus text fill the screen
So that it can be seen easily from a distance.

This is essentially an estimate, using an estimated font aspect ratio
to calculate an approximate font size that can be used that would almost
fill the screen without overflowing it. Unfortunately, this doesn't
quite work - the text fills slightly less of the screen as its length
increases. To compensate for this, there is a slight correction
based on the log of the log of the log of the length, so that it is only
increased very slightly as the text length increases.